src/main/webapp/app/entities/type-exams/route/type-exams-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITypeExams, TypeExams } from '../type-exams.model';
import { TypeExamsService } from '../service/type-exams.service';

import { TypeExamsRoutingResolveService } from './type-exams-routing-resolve.service';

describe('TypeExams routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TypeExamsRoutingResolveService;
  let service: TypeExamsService;
  let resultTypeExams: ITypeExams | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TypeExamsRoutingResolveService);
    service = TestBed.inject(TypeExamsService);
    resultTypeExams = undefined;
  });

  describe('resolve', () => {
    it('should return ITypeExams returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeExams = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTypeExams).toEqual({ id: 123 });
    });

    it('should return new ITypeExams if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeExams = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTypeExams).toEqual(new TypeExams());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as TypeExams })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeExams = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTypeExams).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
